Move ShapeOverlayManager to a separate file
